src/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
click==8.1.7
fastapi==0.103.1
fastapi-cache2==0.2.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
pendulum==2.1.2
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytzdata==2020.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
yarl==1.9.2
